Refactored all views into sub views for easier maintenance
